Milestone1
The program emulates the Snake and ladders game, its operation is very simple, the user enters all the parameters,  such as the number of rows and columns, the number of players, the number of snakes and the number of ladders.  You can make use of commands to effectively use the program, "although
we name it in this section, the game is very intuitive",  remember that the
rules and requirements are better explained in the functional requirements.

The language used is object-oriented java,
functional for Windows, Linux and Mac OS computer operating systems.